Moved export code out of model classes
