+++ src/OpenApi/Export/IOpenApiWriter.cs
@@ -1,0 +1,9 @@
+﻿
+namespace Tavis.OpenApi
+{
+    using System.IO;
+    using Tavis.OpenApi.Model;
+    public interface IOpenApiWriter {
+        void Write(Stream stream, OpenApiDocument document);
+    }
+}
