+++ src/OpenApi/Model/Contact.cs
@@ -4,7 +4,7 @@ using SharpYaml.Serialization;
 
 namespace Tavis.OpenApi.Model
 {
-    public class Contact : IModel
+    public class Contact 
     {
         public string Name { get; set; }
         public Uri Url { get; set; }
@@ -12,17 +12,5 @@ namespace Tavis.OpenApi.Model
         public Dictionary<string, AnyNode> Extensions { get; set; } = new Dictionary<string, AnyNode>();
 
         
-        void IModel.Write(IParseNodeWriter writer)
-        {
-
-            writer.WriteStartMap();
-
-            writer.WriteStringProperty("name", Name);
-            writer.WriteStringProperty("url", Url?.OriginalString);
-            writer.WriteStringProperty("email", Email);
-
-            writer.WriteEndMap();
-        }
-
     }
 }
