+++ src/OpenApi/Model/Header.cs
@@ -7,9 +7,9 @@ using System.Threading.Tasks;
 namespace Tavis.OpenApi.Model
 {
 
-    public class Header : IModel
+    public class Header : IReference
     {
-
+        public OpenApiReference Pointer { get; set; }
         public string Description { get; set; }
         public bool Required { get; set; }
         public bool Deprecated { get; set; }
@@ -23,24 +23,6 @@ namespace Tavis.OpenApi.Model
         public Dictionary<string, MediaType> Content { get; set; }
 
         public Dictionary<string, AnyNode> Extensions { get; set; }
-
-        void IModel.Write(IParseNodeWriter writer)
-        {
-            writer.WriteStartMap();
-            writer.WriteStringProperty("description", Description);
-            writer.WriteBoolProperty("required", Required,false);
-            writer.WriteBoolProperty("deprecated", Deprecated,false);
-            writer.WriteBoolProperty("allowEmptyValue", AllowEmptyValue,false);
-            writer.WriteStringProperty("style", Style);
-            writer.WriteBoolProperty("explode", Explode,false);
-            writer.WriteBoolProperty("allowReserved", AllowReserved,false);
-            writer.WriteObject("schema", Schema, ModelHelper.Write);
-            writer.WriteList("examples", Examples, AnyNode.Write);
-            writer.WriteObject("example", Example, AnyNode.Write);
-            writer.WriteMap("content", Content, ModelHelper.Write);
-
-            writer.WriteEndMap();
-        }
         
     }
 }
