+++ src/OpenApi/Model/Link.cs
@@ -7,27 +7,19 @@ using System.Threading.Tasks;
 
 namespace Tavis.OpenApi.Model
 {
-    public class Link : IModel, IReference
+    public class Link :  IReference
     {
         public string Href { get; set; }
         public string OperationId { get; set; }
-        public Dictionary<string, VariableExpression> Parameters { get; set; }
-        public Dictionary<string, Header> Headers { get; set; }
+        public Dictionary<string, RuntimeExpression> Parameters { get; set; }
+        public RuntimeExpression RequestBody { get; set; }
+
         public string Description { get; set; }
         public Dictionary<string, string> Extensions { get; set; }
 
         public OpenApiReference Pointer { get; set; }
 
 
-        void IModel.Write(IParseNodeWriter writer)
-        {
-            writer.WriteStartMap();
-            writer.WriteStringProperty("href", Href);
-            writer.WriteStringProperty("operationId", OperationId);
-            writer.WriteMap("parameters", Parameters, (w, x) => { w.WriteValue(x.ToString()); });
-
-            writer.WriteEndMap();
-        }
         
     }
 }
