+++ src/OpenApi/Model/Paths.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Tavis.OpenApi.Model
 {
-    public class Paths : IModel
+    public class Paths 
     {
         public IDictionary<string, PathItem> PathItems { get; set; } = new Dictionary<string, PathItem>();
         public Dictionary<string, AnyNode> Extensions { get; set; } = new Dictionary<string, AnyNode>();
@@ -15,17 +15,6 @@ namespace Tavis.OpenApi.Model
         internal void Validate(List<string> errors)
         {
             //TODO:
-        }
-
-        void IModel.Write(IParseNodeWriter writer)
-        {
-            writer.WriteStartMap();
-            foreach (var pathItem in PathItems)
-            {
-                writer.WritePropertyName(pathItem.Key);
-                pathItem.Value.Write(writer);
-            }
-            writer.WriteEndMap();
         }
 
     }
