+++ src/OpenApi/Model/Server.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 namespace Tavis.OpenApi.Model
 {
 
-    public class Server : IModel
+    public class Server 
     {
         public string Description { get; set; }
         public string Url { get; set; }
@@ -12,16 +12,6 @@ namespace Tavis.OpenApi.Model
 
         public Dictionary<string, string> Extensions { get; set; } = new Dictionary<string, string>();
 
-        void IModel.Write(IParseNodeWriter writer)
-        {
-            writer.WriteStartMap();
-
-            writer.WriteStringProperty("url", Url);
-            writer.WriteStringProperty("description", Description);
-
-            writer.WriteMap("variables", Variables, ModelHelper.Write);
-            writer.WriteEndMap();
-        }
         
     }
